+++ paper_uploads/exceptions.py
@@ -1,6 +1,22 @@
 import warnings
 
 
+class CollectionModelNotFoundError(Exception):
+    """
+    Исключение, вызываемое методом `get_collection_class()` элемента коллекции,
+    когда ContentType коллекции существует, но модель коллекции удалена.
+    """
+    pass
+
+
+class CollectionItemNotFoundError(Exception):
+    """
+    Исключение, вызываемое методом `get_item_type_field()` элемента коллекции,
+    когда к классе коллекции не нашлось соответствующего поля для текущего элемента.
+    """
+    pass
+
+
 class InvalidParameter(Exception):
     def __init__(self, value):
         self.value = value